% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_md_imgs.R
\name{bbox_md_imgs}
\alias{bbox_md_imgs}
\title{Diese Funktion ermittelt das Hauptverzeichnis mehrerer Pfade und legt dort eine neue Unterordner an. In diesen Ordner wird die Ordnerstruktur des Hauptverzeichnis mit allen Unterordnern (ohne den neuen Unterordner) kopiert.}
\usage{
bbox_md_imgs(
  md_img_info,
  out.file = tempfile("bbox_img", fileext = ".png"),
  scale = "x480",
  safe.mode = TRUE,
  skip = FALSE
)
}
\arguments{
\item{md_img_info}{data.frame, Datentabelle mit information über Speicherort des Bildes sowie der Geometrie der Bounding Box}

\item{img.in.file}{character, aktueller Speicherort des Bildes (Standardwert ist die Angabe in md_img_info)}

\item{img.out.file}{character, Ausgabepfad des prozessierten Bildes (Standardwert ist die Angabe in md_img_info)}

\item{radius}{numeric, Radius des Rauschfilters}

\item{sigma}{numeric, Stärke des Rauschfilters}
}
\value{
list, enthälte Informationen für die Wiedergabe in R-Shiny (Pfad, Größe)
}
\description{
Diese Funktion ermittelt das Hauptverzeichnis mehrerer Pfade und legt dort eine neue Unterordner an. In diesen Ordner wird die Ordnerstruktur des Hauptverzeichnis mit allen Unterordnern (ohne den neuen Unterordner) kopiert.
}
