% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStandorteNamesCorr.R
\name{checkStandorteNamesCorr}
\alias{checkStandorteNamesCorr}
\title{Führt manuelle Korrektur von Spaltenbezeichnungen der Standorte durch}
\usage{
checkStandorteNamesCorr(
  standorte_import_new_names_corr_auto_sf,
  standorte_names_corr_manual_df
)
}
\arguments{
\item{standorte_import_new_names_corr_auto_sf}{simple feature (sf), Standorte mit automatisch korrigierten Spaltenbezeichnungen}

\item{standort_names_corr_manual_df}{data.frame, Tabelle mit alten und korrigierten Spaltenbezeichnungen}
}
\value{
simple feature (sf), Standorte mit automatisch und manuell korrigierten Spaltenbezeichnungen
}
\description{
Führt manuelle Korrektur von Spaltenbezeichnungen der Standorte durch
}
