% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStandorteNamesWrong.R
\name{checkStandorteNamesWrong}
\alias{checkStandorteNamesWrong}
\title{Führt automatische Korrektur von Spaltenbezeichnungen der Standorte durch}
\usage{
checkStandorteNamesWrong(standorte_import_new_sf)
}
\arguments{
\item{standorte_import_new_sf}{simple feature (sf), Standorte ohne Korrektur der Spaltenbezeichnungen}
}
\value{
simple feature (sf), Standorte mit automatisch korrigierten Spaltenbezeichnungen
}
\description{
Führt automatische Korrektur von Spaltenbezeichnungen der Standorte durch
}
