% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_md_out.R
\name{cleanup_md_out}
\alias{cleanup_md_out}
\title{Entfernt nicht mehr existente Bilder aus mit Megadetector erzeugter JSON}
\usage{
cleanup_md_out(file)
}
\arguments{
\item{file}{Pfad zur mit Megadetector erzeugten JSON-Datei}
}
\description{
Entfernt nicht mehr existente Bilder aus mit Megadetector erzeugter JSON
}
