% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_dir.R
\name{copy_dir}
\alias{copy_dir}
\title{Funktion um Ordnerstruktur in ein neues Verzeichnis zu kopieren}
\usage{
copy_dir(from = choose.dir(), to = choose.dir())
}
\arguments{
\item{from}{character, Pfad des Originalverzeichnisses}

\item{to}{character, Pfad des Zielverzeichnisses}
}
\description{
Funktion um Ordnerstruktur in ein neues Verzeichnis zu kopieren
}
