% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_md_folder_structure.R
\name{create_md_folder_structure}
\alias{create_md_folder_structure}
\title{Diese Funktion ermittelt das Hauptverzeichnis mehrerer Pfade und legt dort eine neue Unterordner an. In diesen Ordner wird die Ordnerstruktur des Hauptverzeichnis mit allen Unterordnern (ohne den neuen Unterordner) kopiert.}
\usage{
create_md_folder_structure(files, md_dirs = "md_empty")
}
\arguments{
\item{files}{Pfade}
}
\description{
Diese Funktion ermittelt das Hauptverzeichnis mehrerer Pfade und legt dort eine neue Unterordner an. In diesen Ordner wird die Ordnerstruktur des Hauptverzeichnis mit allen Unterordnern (ohne den neuen Unterordner) kopiert.
}
