% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{dbConnection}
\alias{dbConnection}
\title{Baut eine Verbindung zur Fotofallendatenbank auf (PostgreSQL).}
\usage{
dbConnection()
}
\value{
Formal class PostgreSQLConnection
}
\description{
Baut eine Verbindung zur Fotofallendatenbank auf (PostgreSQL).
}
