% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectionWorking.R
\name{dbConnectionWorking}
\alias{dbConnectionWorking}
\title{Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu 端berpr端fen.}
\usage{
dbConnectionWorking(psw)
}
\value{
Boolean
}
\description{
Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu 端berpr端fen.
}
\examples{
#'
# returns FALSE
dbConnectionWorking("falsches_passwort")


}
