% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_md_data.R
\name{download_md_data}
\alias{download_md_data}
\title{Download Megadetectorergebnisse
Lade die Ergebnisse der Megadetector-Klassifikation aus der Datenbank}
\usage{
download_md_data(deployments, include.classified = FALSE)
}
\arguments{
\item{include.classified}{boolean, gibt an ob auch bereits klassifizierte Bounding Boxes geladen werden sollen}

\item{directories}{character vector, Pfade zu den gewählten directories, die klassifiziert werden}
}
\description{
Download Megadetectorergebnisse
Lade die Ergebnisse der Megadetector-Klassifikation aus der Datenbank
}
