% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_image_info.R
\name{exif_image_info}
\alias{exif_image_info}
\title{Extrahiere Pfad und Erstellungsdatum der Bilder und gib einen data.frame aus}
\usage{
exif_image_info(
  md_dir,
  info_out = tempfile("out", tmpdir = here::here(), fileext = ".txt"),
  optional_tags = NULL
)
}
\arguments{
\item{md_dir}{character, Verzeichnis in dem rekursiv Bilder gesucht werden}

\item{info_out}{character, Verzeichnis in dem die Tabelle zwischengespeichert wird}

\item{optional_tags}{character, Weitere exiftool-Tags die in der Tabelle eingetragen werden sollen (z.B: "-Model" um das Kameramodell zu extrahieren)}
}
\description{
Prinzipiell lassen sich bei Bedarf mit dieser
}
\examples{
#Checks all images in the current directory recursively
exif_image_info(here::here(), optional_tags = "-Model")

}
