% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_md_info.R
\name{extract_md_info}
\alias{extract_md_info}
\title{Extrahiere Infos aus geladener JSON}
\usage{
extract_md_info(img)
}
\arguments{
\item{img}{data.frame, "Images" Attribute eines einzelnen Bilders aus der mittels Megadetector erzeugten JSON-Datei (z.B. "md_out.json") mit Informationen zur Klassifikation der einzelnen Bounding Boxen sowie deren Geometrie}
}
\value{
data.frame
}
\description{
Extrahiere Infos aus geladener JSON
}
