% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoDB.R
\name{getInfoDB}
\alias{getInfoDB}
\title{Ruft Infos aus Information_Schema ab, um bspw. Spaltennamen einer Tabelle aus der Datenbank abzurufen}
\usage{
getInfoDB(
  type = "column_name",
  schema = "fotofallen",
  table = "fotofallen_standorte_import"
)
}
\arguments{
\item{type}{character, bestimmt welche Information abgefragt werden soll}

\item{table}{character, bestimmt den Namen der Tabelle}

\item{data}{character, bestimmt in welchem Schema die Tabelle vorliegt}
}
\value{
named character vector
}
\description{
Ruft Infos aus Information_Schema ab, um bspw. Spaltennamen einer Tabelle aus der Datenbank abzurufen
}
\examples{
getInfoDB(type = "column_name", schema = "fotofallen", table = "fotofallen_standorte_import")

}
