% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStandorte.R
\name{importStandorte}
\alias{importStandorte}
\title{Ruft Fotofallenstandorte aus der Datenbank ab und lädt zugleich neue Standorte in die akutelle Session ein}
\usage{
importStandorte(file_path, project, target_crs = 25832)
}
\arguments{
\item{file_path}{character, Pfad zu den lokalen Geodaten}

\item{project}{character, Projektbezeichnung der importierten Standorte}

\item{target_crs}{numeric, EPSG-Nummer der Koordinatenreferenzsystems (CRS)}
}
\value{
list, Liste mit neu importierte Standorten, Standorten mit der gegebenen Projektbezeichnung in der Datenbank (fotofallen.fotofallen_standorte) und alle Standorten im Importbereich der Datenbank (import.fotofallen_standorte)
}
\description{
Ruft Fotofallenstandorte aus der Datenbank ab und lädt zugleich neue Standorte in die akutelle Session ein
}
