% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_md_out_imgs.R
\name{load_md_out_imgs}
\alias{load_md_out_imgs}
\title{Lade JSON Datei}
\usage{
load_md_out_imgs(file.path)
}
\arguments{
\item{img}{file.path, Pfad zu der mittels Megadetector erzeugten JSON-Datei (z.B. "md_out.json")}
}
\value{
list, enthält die Images-Attribute der Megadetector-JSON Datei
}
\description{
Lade JSON Datei
}
