% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_md_img2folder.R
\name{move_md_img2folder}
\alias{move_md_img2folder}
\title{Verschiebt Dateien zwischen Verzeichnissen}
\usage{
move_md_img2folder(screened_images, reverse, original_file_col, md_file_col)
}
\arguments{
\item{screened_images}{data.frame, muss Pfade mit Start und Zielverzeichnis enthalten}

\item{reverse}{boolean, in umgekehrter Richtung verschieben (nur zu Testzwecken)}

\item{original_file_col}{character, Pfad der Datei}

\item{md_file_col}{character, Zielpfad}
}
\value{
boolean
}
\description{
Verschiebt Dateien zwischen Verzeichnissen
}
