% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectsDB.R
\name{projectsDB}
\alias{projectsDB}
\title{Gibt alle Projekte in alphabetischer Reihenfolge in fotofallen.projekte in der Datenbank an}
\usage{
projectsDB()
}
\value{
character vector
}
\description{
Gibt alle Projekte in alphabetischer Reihenfolge in fotofallen.projekte in der Datenbank an
}
\examples{
#' projectsDB()

}
