% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_md_classified_data.R
\name{upload_md_classified_data}
\alias{upload_md_classified_data}
\title{Upload Megadetectorergebnisse
Lade die Ergebnisse der manuellen Klassifikation in die Datenbank hoch}
\usage{
upload_md_classified_data(md_classified_df, include.classified)
}
\arguments{
\item{md_classified_df}{data.frame, Datentabelle die mit ShinyFFM2 generiert wird}

\item{include.classified}{boolean, gibt an ob auch bereits klassifizierte Daten hochgeladen werden}
}
\description{
Upload Megadetectorergebnisse
Lade die Ergebnisse der manuellen Klassifikation in die Datenbank hoch
}
