% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_md_data.R
\name{upload_md_data}
\alias{upload_md_data}
\title{Upload Megadetectorergebnisse
Lade die Ergebnisse der Megadetector-Klassifikation in die Datenbank}
\usage{
upload_md_data(md_out_events)
}
\arguments{
\item{md_out_events}{data.frame, Datentabelle mit den Attributen c("file", "conf", "category", "x_off", "y_off", "width", "height", "category_word", "datetimeoriginal", "deployment", "event", "event_id"), die durch die ShinyApp DMCrAI generiert werden}
}
\description{
Upload Megadetectorergebnisse
Lade die Ergebnisse der Megadetector-Klassifikation in die Datenbank
}
