% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadStandorte.R
\name{uploadStandorte}
\alias{uploadStandorte}
\title{Standorte der Session werden in die FVA-Fotofallendatenbank (fotofallen.fotofallen_standorte_import) hochgeladen (PostgreSQL)}
\usage{
uploadStandorte(upload_sf, db_table = "fotofallen_standorte")
}
\arguments{
\item{upload_sf}{simple feature Objekt, Fotofallenstandorte mit korrigierten bzw. überprüften Spaltenbezeichnungen}

\item{db_table}{character, Zieltabelle im Schema fotofallen (entweder "fotofallen_standorte" oder "fotofallen_standorte_import")}
}
\value{
character, Nachricht über den Verlauf des Uploads
}
\description{
Standorte der Session werden in die FVA-Fotofallendatenbank (fotofallen.fotofallen_standorte_import) hochgeladen (PostgreSQL)
}
