% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_md.R
\name{setup_md}
\alias{setup_md}
\title{Setze die Pfade für die Ausführung von Megadetector}
\usage{
setup_md(
  test_md = TRUE,
  md_model_loc = choose.files(caption = "select MD-model file, e.g. md_v5a.0.0.pt", multi
    = FALSE),
  py_scripts_loc = choose.dir(caption =
    "select directory where ai4eutils, cameratraps and yolov5 are located - path may not contain blanks!")
)
}
\arguments{
\item{test_md}{boolean, gibt an, ob nach dem Setzen der Pfade ein Testlauf durchgeführt werden soll}

\item{md_model_loc, }{character, Verzeichnis in dem Megadetector-Modell liegt}

\item{py_scripts_loc}{character, Verzeichnis in dem die für Megadetector Python-Skripte hinterlegt sind
#'}
}
\description{
Setze die Pfade für die Ausführung von Megadetector
}
